Mostly update example, fix a issue in physics as wel..

+++ src/com/moka/core/time/Time.java
@@ -24,9 +24,9 @@ public class Time extends SubEngine
 		}
 	}
 
-	public double getDelta()
+	public float getDelta()
 	{
-		return delta;
+		return (float) delta;
 	}
 
 	public double getElapsed()
